+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to casper-client-sdk.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.32
+
+### Added
+
+- `CasperServiceByJsonRPC.getBlockInfoByHeight(height)`
+
+### Fixed
+
+- `CasperServiceByJsonRPC.getBlockInfo(hash)` to return requested block, not the last one.
+
 ## 1.0.25
 
 ### Added